Modify route:popular, for db sechema change
