+++ representers/food_ranks.rb
@@ -1,12 +1,9 @@
+# frozen_string_literal: true
+require_relative 'food_rank'
 
+# Represents overall group information for JSON API output
+class FoodRanksRepresenter < Roar::Decorator
+  include Roar::JSON
 
-# Search query for postings in a group by optional keywords
-class Food_RanksRepresenter
-  def self.call(results)
-    output = '['
-    results.each {|a|
-      output = output + {food_id: a.food_id, name: a.name, tag_amount: a.tag_amount, rank: a.rank}.to_s + ','}
-    output.chop!
-    output = output + ']'
-  end
+  collection :foods, extend: FoodRankRepresenter, class: Food
 end
