+++ services/get_popular.rb
@@ -6,17 +6,11 @@ class GetPopular
 
   def self.call(criteria)
     popular = FindRankFoodQuery.call(criteria)
-    result = []
-    for count in 0..popular.length-1
-      foodrank = popular[count]
-      food = Food.find(id: foodrank.food_id)
-      result[count] = RankMaterial.new(photo:food.photo,food_id:food.id,name:food.food_name,tag_amount:foodrank.tag_amount,rank:count+1)
-    end
 
-    if (result).nil?
+    if (popular).nil?
       Left(Error.new(:not_found, 'Error in GetLessPopular'))
     else
-      Right(result)
+      Right(Foods.new(popular))
     end
   end
 end
