+++ values/foods.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+Foods = Struct.new :foods
